Updated Stock.java

+++ OOPT/src/main/java/com/example/javafx_login/classes/Stock.java
@@ -1,6 +1,7 @@
 package com.example.javafx_login.classes;
 
 import java.io.*;
+
 public class Stock {
     //Variables
     private static int totalProducts;
@@ -10,42 +11,37 @@ public class Stock {
         return totalProducts;
     }
 
+    static File file = new File("stock.txt");
+
     //Methods
-    //create file(used for one time)
-    public static void createStockFile(){
-        try {
-            BufferedWriter writer = new BufferedWriter(new FileWriter("stock.txt"));
-            writer.write("Product Name | Serial Number | Stock Quantity | Price (RM)\n");
-            writer.write("==========================================================\n");
-            writer.close();
-        }
-
-        catch (IOException e) {
-            System.out.println("Unable to create file!");
-            e.printStackTrace();
-        }
-    }
-
     //appending/add file data
-    public static void addStock(String name, String serialNum, int quantity, double price){
+    public static void addStock(String name, String serialNum, int quantity, double price) {
+        if (!file.exists()) {
+            try {
+                BufferedWriter writer = new BufferedWriter(new FileWriter("stock.txt"));
+                writer.write("Product Name | Serial Number | Stock Quantity | Price (RM)\n");
+                writer.write("==========================================================\n");
+                writer.close();
+            } catch (IOException e) {
+                System.out.println("Unable to create file!");
+                e.printStackTrace();
+            }
+        }
+
         try {
             BufferedWriter writer = new BufferedWriter(new FileWriter("stock.txt", true));
-            writer.write(name + " | "  + serialNum + " | " + quantity  + " | " +  price + "\n");
+            writer.write(name + " | " + serialNum + " | " + quantity + " | " + price + "\n");
             writer.close();
-        }
-
-        catch (IOException e) {
+        } catch (IOException e) {
             System.out.println("Unable to open file!");
             e.printStackTrace();
-        }
-
-        finally {
+        } finally {
             totalProducts++;
         }
     }
 
     //display all products(regardless availability)
-    public static void getAllProducts(){
+    public static void getAllProducts() {
         try {
             BufferedReader reader = new BufferedReader(new FileReader("stock.txt"));
             String line;
@@ -54,30 +50,27 @@ public class Stock {
                 System.out.println(line);
             }
             reader.close();
-        }
-
-        catch (IOException e){
+        } catch (IOException e) {
             System.out.println("Unable to read file!");
             e.printStackTrace();
         }
     }
 
     //check product availability
-    public static boolean isProductAvailable (String itemName){
+    public static boolean isProductAvailable(String itemName) {
         String[] selectedProductDetails = new String[4];
         try {
             BufferedReader reader = new BufferedReader(new FileReader("stock.txt"));
             String line;
 
             while ((line = reader.readLine()) != null) {
-                if(line.contains(itemName)) {
+                if (line.contains(itemName)) {
                     selectedProductDetails = line.split(" \\| ", 0);
                 }
             }
             reader.close();
-        }
 
-        catch (IOException e){
+        } catch (IOException e) {
             System.out.println("Unable to read file!");
             e.printStackTrace();
         }
@@ -86,20 +79,19 @@ public class Stock {
     }
 
     //display the details of the selected product
-    public static void displayProductDetails (String itemName) {
+    public static void displayProductDetails(String itemName) {
         try {
             BufferedReader reader = new BufferedReader(new FileReader("stock.txt"));
             String line;
             while ((line = reader.readLine()) != null) {
-                if(line.contains(itemName))
+                if (line.contains(itemName))
                     System.out.println(line);
             }
             reader.close();
-        }
-
-        catch (IOException e){
+        } catch (IOException e) {
             System.out.println("Unable to read file!");
             e.printStackTrace();
         }
     }
+
 }
